+++ src/GRA.Domain.Service/ChallengeService.cs
@@ -19,6 +19,7 @@ namespace GRA.Domain.Service
         private readonly IBranchRepository _branchRepository;
         private readonly ICategoryRepository _categoryRepository;
         private readonly IChallengeRepository _challengeRepository;
+        private readonly IChallengeGroupRepository _challengeGroupRepository;
         private readonly IChallengeTaskRepository _challengeTaskRepository;
         private readonly IPathResolver _pathResolver;
         private readonly ITriggerRepository _triggerRepository;
@@ -31,6 +32,7 @@ namespace GRA.Domain.Service
             IBranchRepository branchRepository,
             ICategoryRepository categoryRepository,
             IChallengeRepository challengeRepository,
+            IChallengeGroupRepository challengeGroupRepository,
             IChallengeTaskRepository challengeTaskRepository,
             IPathResolver pathResolver,
             ITriggerRepository triggerRepository,
@@ -41,6 +43,8 @@ namespace GRA.Domain.Service
             _categoryRepository = Require.IsNotNull(categoryRepository, nameof(categoryRepository));
             _challengeRepository = Require.IsNotNull(challengeRepository,
                 nameof(challengeRepository));
+            _challengeGroupRepository = Require.IsNotNull(challengeGroupRepository,
+                nameof(challengeGroupRepository));
             _challengeTaskRepository = Require.IsNotNull(challengeTaskRepository,
                 nameof(challengeTaskRepository));
             _pathResolver = Require.IsNotNull(pathResolver, nameof(pathResolver));
@@ -509,6 +513,121 @@ namespace GRA.Domain.Service
             {
                 System.IO.File.Delete(filePath);
             }
+        }
+
+        public async Task<ChallengeGroup> GetGroupByIdAsync(int id)
+        {
+            var challengeGroup = await _challengeGroupRepository.GetByIdAsync(id);
+            if (challengeGroup == null)
+            {
+                throw new GraException("The request challenge group could not be accessed or does not exist");
+            }
+            await AddBadgeFilenames(challengeGroup.Challenges);
+
+            return challengeGroup;
+        }
+
+        public async Task<ChallengeGroup> GetGroupByStubAsync(string stub)
+        {
+            return await _challengeGroupRepository.GetByStubAsync(GetCurrentSiteId(), stub.ToLower());
+        }
+
+        public async Task<DataWithCount<IEnumerable<ChallengeGroup>>>
+            GetPaginatedGroupListAsync(BaseFilter filter)
+        {
+            VerifyPermission(Permission.ViewAllChallenges);
+            filter.SiteId = GetCurrentSiteId();
+            return new DataWithCount<IEnumerable<ChallengeGroup>>
+            {
+                Data = await _challengeGroupRepository.PageAsync(filter),
+                Count = await _challengeGroupRepository.CountAsync(filter)
+            };
+        }
+
+        public async Task<ServiceResult<ChallengeGroup>> AddGroupAsync(ChallengeGroup challengeGroup,
+            List<int> ChallengeIds)
+        {
+            VerifyPermission(Permission.AddChallengeGroups);
+
+            var siteId = GetCurrentSiteId();
+            var stub = challengeGroup.Stub.Trim().ToLower();
+            var existingStub = await _challengeGroupRepository.StubInUseAsync(siteId, stub);
+            if (existingStub == true)
+            {
+                throw new GraException($"A challenge group with the link {stub} already exists.");
+            }
+
+            var serviceResult = new ServiceResult<ChallengeGroup>();
+            challengeGroup.SiteId = siteId;
+            challengeGroup.Stub = stub;
+
+            var validChallengeIds = await _challengeRepository.ValidateChallengeIdsAsync(siteId,
+                ChallengeIds);
+
+            if (ChallengeIds.Count != validChallengeIds.Count())
+            {
+                serviceResult.Status = ServiceResultStatus.Warning;
+                serviceResult.Message = "One or more of the selected challenges could not be added to this group.";
+            }
+
+            serviceResult.Data = await _challengeGroupRepository.AddSaveAsync(
+                GetClaimId(ClaimType.UserId), challengeGroup, validChallengeIds);
+
+            return serviceResult;
+        }
+
+        public async Task<ServiceResult<ChallengeGroup>> EditGroupAsync(
+            ChallengeGroup challengeGroup, List<int> ChallengeIds)
+        {
+            VerifyPermission(Permission.EditChallengeGroups);
+
+            var siteId = GetCurrentSiteId();
+            var serviceResult = new ServiceResult<ChallengeGroup>();
+            var currentChallengeGroup = await _challengeGroupRepository.GetByIdAsync(
+                challengeGroup.Id);
+            challengeGroup.SiteId = currentChallengeGroup.SiteId;
+            challengeGroup.Stub = currentChallengeGroup.Stub;
+
+            var validChallengeIds = await _challengeRepository.ValidateChallengeIdsAsync(siteId,
+                ChallengeIds);
+            if (ChallengeIds.Count != validChallengeIds.Count())
+            {
+                serviceResult.Status = ServiceResultStatus.Warning;
+                serviceResult.Message = "One or more of the selected challenges could not be added to this group.";
+            }
+
+            var currentChallengeIds = currentChallengeGroup.Challenges.Select(_ => _.Id);
+            var challengesToAdd = ChallengeIds.Except(currentChallengeIds);
+            var challengesToRemove = currentChallengeIds.Except(ChallengeIds);
+
+            serviceResult.Data = await _challengeGroupRepository.UpdateSaveAsync(
+                GetClaimId(ClaimType.UserId), challengeGroup, challengesToAdd, challengesToRemove);
+
+            return serviceResult;
+        }
+
+        public async Task RemoveGroupAsync(int groupId)
+        {
+            VerifyPermission(Permission.EditChallengeGroups);
+            await _challengeGroupRepository.RemoveSaveAsync(GetClaimId(ClaimType.UserId), groupId);
+        }
+
+        public async Task<List<Challenge>> GetByIdsAsync(IEnumerable<int> challengeIds)
+        {
+            VerifyPermission(Permission.ViewAllChallenges);
+            return await _challengeRepository.GetByIdsAsync(GetCurrentSiteId(), challengeIds);
+        }
+
+        public async Task<List<ChallengeGroup>> GetGroupsByChallengeId(int id)
+        {
+            VerifyPermission(Permission.ViewAllChallenges);
+            return await _challengeGroupRepository.GetByChallengeId(GetCurrentSiteId(), id);
+        }
+
+        public async Task<bool> StubInUseAsync(string stub)
+        {
+            VerifyPermission(Permission.AddChallengeGroups);
+            return await _challengeGroupRepository.StubInUseAsync(GetCurrentSiteId(), stub.ToLower());
         }
     }
 }